Refactored node creation code into main.go
node.go now only contains code pertaining to internal node logic.
