+++ src/Mutual_Exclusion/node.go
@@ -140,42 +140,20 @@ func (n *P2PNode) AskAllPeers() {
 	n.peerLock.RUnlock()
 }
 
-func main() {
-	node1 := CreateNode(0, 3)
-	node2 := CreateNode(1, 3)
-	node3 := CreateNode(2, 3)
-
-	// Start the gRPC server
-	go StartServer(node1, ":50051")
-	go StartServer(node2, ":50052")
-	go StartServer(node3, ":50053")
-
-	time.Sleep(3 * time.Second)
-
-	// Add peers (simulate peer discovery for demonstration)
-	node1.AddPeer("localhost:50052")
-	node1.AddPeer("localhost:50053")
-
-	node2.AddPeer("localhost:50051")
-	node2.AddPeer("localhost:50053")
-
-	node3.AddPeer("localhost:50051")
-	node3.AddPeer("localhost:50052")
-
+func (n *P2PNode) Start() {
 	// Simulate sending a message to a peer
-
 	for {
-		node1.Request_Critical = true
+		n.Request_Critical = true
 		log.Printf("Node 1 requests access \n")
-		node1.Our_Timestamp = node1.Highest_Timestamp + 1
-		node1.Outstanding_Reply = node1.N - 1
-		node1.AskAllPeers()
+		n.Our_Timestamp = n.Highest_Timestamp + 1
+		n.Outstanding_Reply = n.N - 1
+		n.AskAllPeers()
 		timeout := 0
 		for {
 			if timeout == 30 {
 				//ReleaseCS()
 			}
-			if node1.Outstanding_Reply == 0 {
+			if n.Outstanding_Reply == 0 {
 				//Enter Critical section
 				//Do something
 				//Exit Critical section
